More Poller work
Take a different route for polling in background tasks.
- Initial thought was using shelly_api.py over a small subset
- It doesn't appear to cost more to make a GetStatus call
  - IIRC this provides the same data as individual getters since
    <SubComponent>.GetStatus returns just that subset

The shelly_api remains solid for scripting and places where values
need to be set, but bulk collection can be smarter. The devices
already support gzipped responses from the REST API, so bandwidth
isn't a concern.

Most of the added work involved flattening the nested dict the API
returns. Upon further thought, this should always happen on receive.
If flattening fails, dump the data into a text file for rejected
entries.